LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/getBoundingClientRect-newline-expected.txt

123456
789012

PASS Range.getBoundingClientRect() should return the first position of the next line when the collapsed range is a newline character
